[Silver III] Title: 주유소, Time: 340 ms, Memory: 35500 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -18,7 +18,7 @@ public class Main {
         for(int i=0; i<N; i++) city[i] = Integer.parseInt(st.nextToken());
 
         long cost = 0;
-        int minPrice = city[0];
+        long minPrice = city[0];
 
         for(int i=0; i<N-1; i++){
             minPrice = Math.min(minPrice, city[i]);
